Integrate nodescan worker into launcher
This checks for connectivity and scans for ssh host keys on new nodes.

Unlike nodepool, we will store the flag indicating whether we should
scan for keys as well as the boot timeout on the node.  This is
because later in the node creation process, we no longer have easy
access to the label with that configuration information (and it may
change or no longer be present by the end).

Semi-related:

There are two locations in the launcher where we create new node
objects; one of them was missing some optional initializers related
to image information.  It is updated so that both are at parity.

Change-Id: I8ff7188cc8c8ae8f33ab1e92f3f5b0b91b9bfd41