Launcher: continue processing nodes during nodescan
This corrects a bug where the launcher will go idle during the nodescan
process if it takes longer than one iteration.  Our tests fake out the
nodescan requests to such a degree that they complete in one pass through
the loop, so doesn't hit this case.  In production, we need to tell the
main loop to keep running until the request is complete.

Due to the nature of the fakes and what nodescan does, this is difficult
to test.

Change-Id: Iacb0942eb91e0fab72b61bc9a41a196612929937