Fix error with empty nodeset request
If a tenant has any providers active, then node requests for empty
nodesets are handled by the new zuul-launcher mechanism rather than
the old nodepool mechanism.  The request still don't actually go
to the launcher, but the internal classes and methods are used to
fulfill an empty "nodeset request" rather than a "node request".

This correctly a typo in a method call that is only used in that
case and adds test case coverage.

Change-Id: I60ce6c0efa91b5b65b3d530450e9bd4abbc22db9