Report some scheduler cpu stats
This reports the overal user/system time for the scheduler, as well
as the cpu/system time for all of the background processing threads
for connections.

This will allow us to understand what percent of the time a scheduler
spends dealing with background connection processing versus layout
updating and the main run handler.

Change-Id: Ifdbd169ba573a3f107f3ea9c5308e57f418975a3