Fix node listing web API
The node "type" field in the web api is intended to be backwards
compatible with nodepool nodes.  They provide a list of strings.
Update zuul-web to provide a list of one string for the type field.
The new "label" field retains its single-list type for future use
once nodepool backwards-compatability is not needed.

This corrects an error in the web ui on the nodes page which expects
the "type" field to be a list.

Change-Id: I1b36d5edba7bd70caa139d634fbd63042ae8dddb