Launcher: fix image decompression
We did not correctly check the image filename extensions when
deciding whether to decompress an image.  Correct that by downloading
to a path with the extension (since zstd will want to see a file
with a .zst extension and then remove it), and then check the extension
of the downloaded path.

Update the tests to exercise this.

Add zstd to bindep so it's installed for the tests and expressed as
a requirement for Zuul in general.  Remove the launcher-specific addition
of zstd to the container image for simplicity.

Change-Id: Ie6ba9c54ed0246157fb0d1ea820b758872e60d98