Launcher: improve nodescan testing
This improves the nodescan testing by:

* Removing the fake handling from the production code
* Adds explicit host-key-checking: false settings to most nodepool
  test fixtures.
* Adding explicit success/failure tests for nodescan with the full
  node request lifecycle.
* Moves the test-only fakes into their own file so they can be shared
  with the new launcher nodescan tests as well as the dedicated
  nodescan unit tests.
* Allows for customizable timeouts in the requestNodes helper method
  since the failing test takes a while.

It also moves the boot-timeout parameter from labels to images, and,
more importantly, allows it to be set on the section object itself.
Typically a boot-timeout would be a characteristic of the cloud (ie,
the cloud is slow to boot anything) or an image (this image is slow
to boot).  It's less likely to be a characteristic of a specific label.

Change-Id: If6361cb437ad28472f154f3af9dde190d43c0d1e