Fix nested provider validation with hyphens
The method we use to flatten a complete provider config and validate
it used the toDict method to re-render the nested label/image/flavor
config objects as dictionaries underneath the resulting provider
config.  However, we want to validate the resulting provider config,
which means we need to use hyphens instead of underscores.

Add a new method that renders the config object back to its original
form for use in flattening.

Change-Id: I7f212b0a824b9c473b17ac85df6dea8bdd61e53c