Fix missing console messages on short tasks
A recent fix to the stream callback removed the delay that happens
when we wait for a connection to a server for a skipped task.

However, if a connection for a real command task is slow, and that
same task is very fast, then it can result in the stream callback
exiting before handling any output.  To correct this, check whether
we are in the grace period before we exit when we are requested to
stop during the connection phase.

If we are handling a skipped task, there will be no grace period and
we can exit immediately.  If we are handling a normal task, we will
have a 10 second grace period, so as long as we are still within that
window, we will process log entries (until the log is complete, which
should be immediately for a normal task).  If the connection attempt
itself exceeds the grace period, we will still not get any data, but
that is expected behavior.

Change-Id: Ibeea5e55f3b0b39191283721d5fce46a7c367ece