Add a synchronous TreeCache option
We developed the TreeCache to efficiently handle a large number
of objects that change frequently.  To do that, we spawn two threads
so that if the system is under heavy load, we create a backlog
of tasks.

We may want to use the TreeCache pattern for trees which are smaller
and rarely change.  Rather than needing to build a system with
watches from scratch, let's add an option to the TreeCache so that
it will run synchronously.  That let's use reuse the watch setup in
TreeCache without needing to spawn two more threads for every cache
that we define.

Change-Id: I9723d12eb67a39ac26d882f1b2a96551036c9cb7