Implement command for deleting OIDC signing keys
When OIDC signing keys are compromised, we need to delete the
old keys and regenerate new keys.

This change adds a command in zuul-admin tool for deleting
OIDC signing keys of an algorithm. New keys would be generated
automatically when needed.

Meanwhile, a `password` is not required in those deleting,
exporting, importing, etc. operations, the checking for password
configurations are removed.

Change-Id: I66be703d642346b5feef10431838b010a52b9065