Make Gerrit event pre-processor multi-threaded
Pre-process Gerrit events in thread pool executor similar to how this is
already done for Github.

NOTE: this commit temporarily skips a failing unit test.
The next commit will fix that test.  These two commits are easier
to understand on their own, thus the unusual step of the temporary skip.

Change-Id: I8b6a02376a5294e691dc27b7f2e38a00130deaff