gerritdriver: Fail early on gerrit 404 results
When encountering an HTTP 404 from gerrit when trying to get a change,
we currently retry the request (3 times, with a 1 sec wait in between)
and will finally jump out of the event handler. This causes an
additional 1 second time penalty until re-election of the event handler.

Instead, return early from event processing when encountering an HTTP
404 and skip only that single event without handler re-election.

However, when getting needed_by changes, we don't want to let that
exception propagate all the way up to skipping the complete event, so we
simply ignore failed fetches of needed_by changes.

Change-Id: I67ad8946dc7ca5a318db3ddf3750802e8c068f1d