Use change networks in gerrit driver queries
We want to use a thread pool executor to process multiple gerrit
events in parallel in order to improve throughput even if gerrit is
slow to respond to queries.  But the gerrit driver is heavily
optimized to reduce the number of queries it makes, which means if we
want to run parallel event processing we need to somehow coordinate
among those threads to avoid repeating queries in different threads.

This change introduces a novel way of avoiding repeated queries from
independent threads.

Let's define a change network as a set of changes that are connected
by dependencies.  With no coordination, when one thread starts
processing an event, it may start to build its change network with
change A; another thread may process a different event and start to
build its network with change B.  If A and B are related by
dependencies, the two threads will both eventually perform queries for
both changes.  Thread 1 querying A then B, and thread 2 querying B
then A.

This change introduces coordination by having each thread record its
change network as it builds it out.  These data are recorded in a
ChangeNetworkFuture.  Before performing a query for any change, the
thread will ask a central coordinator (the ChangeNetworkManager)
whether it should proceed with the query.  The ChangeNetworkManager
knows all of the current change network graphs in operation.  If the
thread's network graph intersects another currently active graph, the
ChangeNetworkManager declares a conflict and determines the winner
based on which thread started its work first.  If that is the calling
thread, then it proceeds with the query.  If it is not the calling
thread, then it aborts its work and unwinds the stack all the way back
to its first query.  It then waits until the thread that did win the
conflict finishes, and once it has, it starts its work over again.

Because a losing thread may still have performed queries that a
winning thread did not, the results of every query are cached on the
ChangeNetworkFuture, and losing networks are merged with winning
networks.  Therefore, in our example above, the following sequence
would result:

[1] Check change cache for change A; not present
[1] Create network X
[1] Request permission to query change A
[1] Query change A, add results to network X
[2] Check change cache for change B; not present
[2] Create network Y
[2] Request permission to query change B
[2] Query change B, add results to network Y
[1] Request permission to query change B
[1] Conflict detected; this thread wins
[1] Merge network Y results into network X
[1] Use cache results instead of querying B
[2] Request permission to query change A
[2] Conflict detected; this thread loses
[2] Unwind call stack to initial query
[2] Wait for network X to finish
[1] Store change B results in ZK
[1] Store change A results in ZK
[1] Mark network X finished
[2] Resume since network X is finished
[2] Check change cache for change B; up to date

Change-Id: Ie741f03a9e1070de91c689ffad0a3a909cdc18df