Change network future collision resolution
Update the network future collision resolution so that the first
network to query a change always wins, rather than the oldest.

In pratice, they will often be the same, but using this resolution
method may avoid some unecessary writes to the ZK change cache in
some cases.  With this change, we know at least that only one of
the threads could possbily write the conflicting change to ZK.
In other words, we stop the possibility of ZK change cache writes
as quickly as possible.

Change-Id: Ib9d95c51fc3eb392c6213228efe6aa1e00244a9b