Copy query cache from winning network
We merge the query cache from the losing network into the winning
network so the winner can benefit from any queries the loser performed
but the winner did not.

After the winner completes, in most cases the loser may not need to
do any work, and that's the end of it.  But if the loser decides that
the change updates made by the winner are too old for it, it may need
to walk the network again.  In that case, let's copy the query cache
from the winner to the new, restarted, network of the loser, so that
if the winner did happen to perform any queries sufficiently recently,
we can still benefit from those.

Change-Id: Icd370f79ca1115934520d02a03387b7b4d502f60