Pass event id when getting topic dependencies
The initial goal of this change was to  prevents error logs like "Change
<ChangeKey ...> not found in cache and no event" and annotates related
log messages with the correct event id.

Since we no longer populate a change based on the event, we can remove
that log altogether. The Pagure driver was still using the change URL
from the event, but it looks like that could already have lead to a
wrong change URL in case of non-topic dependencies so this was removed.

Change-Id: I6980bdebc9caf5a81a7833f01550e3e7e92d492a