Launcher: delete image files in case of error
After we attempt to decompress a .zst image, we should delete
the original.  Typically the decompression progrem should do that
for us if it succeeds, but if it doesn't, then we should do it
ourselves.  In no case do we want the original after this point.

Additionally, if we encounter any error during the initial download,
remove the partially completed file.

Change-Id: I8d89074b1f1395ddafd751bbbddd8c959634643a