Make connection event iteration more efficient
With the multi-threaded event connector we could run into a situation
where we'd iterate over pending events multiple times. This is
especially the case when there are a lot of pending events and the TPE
is at it's max. processing capacity. This lead to events being
deserialized on each loop iteration.

To make this more efficient we'll pass the highest event ID that's
currently still processing as an offset and only return events that have
an ID larger than that.

Change-Id: I363f004fe3cbc4586a0ac0d44831932ae58210d6