Process gerrit change-merged events before ref-updated
Most Zuul drivers provide file information on push events, but
Gerrit does not.  Indeed, it doesn't even provide file information
on change events.  Most Zuul drivers trigger configuration updates
when zuul.yaml changes on push events, but the Gerrit driver uses
change-merged because of this limitation.

Because the ref-updated (push) event arrives before the change-merged
event, that means that post pipelines may run with the old
configuration.

To correct this we have some options:

1) Re-order the events in the order that would be more convenient.
2) Figure out the change associated with a ref-updated event and
   query it for a file list and attach that to ref-updated.
3) Emit a synthetic internal event to perform reconfiguration.
4) Perform a git diff on ref-updated events to get a file list.

Option #1 will solve the immediate issue, but we must ensure that
we aren't violating any internal assumptions about event order,
or that changing the event order won't cause problems for users.

Option #2 sounds ideal because it brings the gerrit driver into
parity with the other drivers.  However, that would mean that jobs
in post pipelines, which may be running now because they do not
have a files list, may no longer run once the fil list is present
on the Branch object.  Additionally, it is not possible to associate
change-merged events with ref-updated events when using the
(deprecated) checks plugin, so the old behavior would need to remain
in place for that case.

That may still be the ideal solution.  If desired, we can build on
this change (because the mechanism to associate the events will be
the same) along with the necessary release notes (and possible version
bump).  Meanwhile, we can at least fix the reconfiguration issue
with this change and validate the event association mechanism.

Option #3 would allow us to thread the needle and solve the
reconfiguration order issue without needing to address the files
issue of option #2 or the event order of #1.  It still requires
event association, and so is just as complex as #1, and maybe moreso
since we'll be making a new type of event.

Option #4 is the most correct change, but we would need a merger
to perform work during early event ingestion or processing, which
is a line we are currently not willing to cross due to the severe
latency.

This change implements option #1 since it is relatively simple and
minimal, compared to the others.  It comes with a recommendation to
consider Option #2 in the future as well.

To avoid potential internal issues with the change in order, we
use the older ltime on both changes (so that we don't have ltime
going backwards).  As for user-visible ordering, since change-merged
and ref-updated changes are usually handled by different pipelines,
in any multi-scheduler system it is a race to begin executing jobs,
so most users should not notice post jobs now running after publish
jobs.

Inspired by Joseph Kostreva's change https://review.opendev.org/924079

Change-Id: I4f54f8c780cd837835f5638fbe19ed1d152c10cb