Fix unit test fd leak
When we imported the nodescan worker, we copied a mistake where we
create two pipe file descriptors and never close them.  This does
not affect production, but they are leaked during unit tests and
can cause problems with the os ulimit.

We previously corrected a similar problem in the zuul web server,
but never applied the same fix to nodepool (likely due to its
lighter unit tests not forcing the issue).

Fix the nodescan worker and add comments to both locations so
that if we copy the pattern again, hopefully we will remember
to copy the fix.

Change-Id: Ia54ee21d88ab9e77b7ba7ec2060786391c1875af