Process semaphore release as tenant events
Processing semaphore release events and distributing pipeline semaphore
released events to other pipelines is slow and can take a
non-significant amount of time.

Since this is more of a tenant-level responsibility we'll dispatch the
event to the tenant management event queue which will take care of
distributing the event to all the pipelines.

The code path for handling this event as part of the pipeline result
event queue is kept for backward compatibility and can be removed at a
later time.

Change-Id: Ief91356aa01a282e25e51947e93eee3412b90b69