Only notify pipeline once about semaphore release
We've seen a large number of semaphore release events being processed
each leading to multiple pipeline semaphore release events for the same
pipeline. Since the pipeline semaphore release event is just a
notification and doesn't carry any information it is enough to notify
the pipeline once.

To achieve this we track pipelines that have been notified of a
semaphore release and don't send additional pipeline semaphore release
events in the same iteration of the tenant management event queue.

Change-Id: Idedaf5496430a218a1c4b79fdc3b4c58824a7da9