Update patch set 18
Patch Set 18: Code-Review-1

I don't think we should put a string in the log field as then the web ui or other tools may try to fetch that string as a url (or otherwise would need complex handling to deal with it).

Instead, I think we should truncate it, or omit it entirely (set it to null, or the empty string).

It would be okay to add a warning message to the build message string.

Note that the actual issue should be fixed by the zuul-jobs change, so I agree that the rest of this change is desirable in case something goes wrong we still retain some information about the build, but we don't have to worry too much about providing a message since it shouldn't happen.

Patch-set: 18
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1, fb4524536f3acf0b1d7061d4b54694924caaf7c6