Fix race in test_fail_fast_no_cancel_completed
This tests wants the result events to be sequenced in a specific
order, but because the result event is emitted after the build
is removed from our list of running builds and the event is emitted
in the AnsibleJob worker thread, it is possible for the events to
arrive in the wrong order.

Correct this by waiting for each AnsibleJob thread to complete.

Change-Id: I9ab2022eea42ea597b2e2d72e78b6cdcfd7a1ad9