Improve nodescan worker cleanup
Rather than relying on __del__ to close the pipe fds, let's use
the join() method which we should be calling during our shutdown
procedure anyway.  This will ensure the pipes are closed even
if the object is not GC'd.

Also, remove the backref from the request to the worker to increase
the chances that the object can be GC'd.

Change-Id: I71eb34b2502adebe7b3bf32eefb0fa540535f8f7