Store failed provider node data on request
Currently we keep track of failed node builds by retaining a list
of uuids for each node we have assigned to a label.  When we want
to figure out how many failures we have had, we look up the nodes
in that list.  That means we need to keep the node records in zk
for the duration of the request (even failed nodes).  We do the
same for quota failures, and they are unbounded, meaning we could
end up with infinite zk storage requirements.

This change adjusts how we store that information as follows:

Instead of storing a list of failed and successful nodes for each
label in the request, we will store the uuid of only the latest
node for each label.  The request will forget about previous nodes,
and they can be immediately deleted from zk.

The request will store a list of providers which have had failed
requests for each label position.  We will not store quota failures
in this list, only failures which count against the provider attempts
counter.

These changes are encapsulated in the change from the "provider_nodes"
attribute to the "provider_node_data" attribute on the request.

Since this requires an adjustment to the provider selection algorithm,
we will flesh it out a bit more.  At a high level, it now attempts to
do the following:

* Pick a single provider for every node (or most nodes) in the request
  if possible.  This facilitates multinode jobs that want LAN networking.
* Otherwise, pick the provider with the most available quota for each
  label that has not exceeded the number of allowed attempts.
* Randomize among those providers for different requests.

We also run the delete state machine for a tempfailed node in case
a provider may have started creating resources before encountering
the quota.  We should start to unwind those resources in that case.

Change-Id: Ife5a1d1c313a3f2f5e6132f269e1e3a0b6be0977