Delete node records as soon as practical
Since the request handler no longer relies on old node records in
ZK, we can delete them as soon as the request is done with them.

The request thread needs to know that it's node cache is up to date,
and it does that by verifying that every node in the requset appears
in the node cache.  That means we must not delete node records
as long as they are referenced by a request.

The node thread does not currently know if its request is up to date,
and it must know that in order to know if it's safe to delete a node
record.  It isn't enough to say that the node uuid does not appear
in the request node list because if the request cache is out of date,
the node thread could be operating on data *before* the node was added
to the request.  In other words, before we can decide if it is safe
to remove a node based on it not being in a request, we must ensure
that the request is at least new enough to have had the node in it in
the first place.

To accomplish that, we store the zstat of the node request where the
node was added.  We don't actually know what that is since we write
the node request after the node, so we add 1 to the current zstat.

Change-Id: I83c6a54e034c2629133f2222c7176af9de6b8d23