WIP: Attempt to avoid quota starvation
This attempts to avoid quota starvation by turning our distributed
system of parallel provider processors into a serialized system
as we approach quota.

After 95% quota utilization, the system will ensure that the oldest
outstanding node for a provider begins its state machine before the
next will start.

If we want to try this out, we will still need to figure out how
to test it; it may be unreasonably difficult.

Change-Id: I24a7ffc392d8a936b9f76c7f3d8879e6f50c0573