Respect replication delay in event pre-processing
With Ie741f03a9e1070de91c689ffad0a3a909cdc18df we considered a Gerrit
change valid for a given event if it's ltime was newer than the
connection event ltime.

However, this can cause issues as the Gerrit replication delay is not
considered in that case.

To fix this we'll get the current ltime after the delay and pass this as
the minimum ltime required for a change.

Change-Id: Ib02d9cde2a6f5f1cffd744a92b8f451eb58a3879