Add test for locked branch filtering on validation
Change Ia4992e55d1f0448d98667cbfd71ba033c966918e fixed a bug with
performing tenant-validation on locked branches.  This adds a test
which exercises that code path.

The validation test in this change is effective.  The startup test
did not fail even without the bugfix (because the initial call to
fetch the branch list does not result in a cat job).

Change-Id: I2fbcf76df338fba80adf954ac066a6e63111f3b2