Pin confluent-kafka
Building this wheel is failing in opendev.

Disable the upgrade job because we currently can't install the
old-side deps.  It will be re-enabled in the next commit.

Increase the delay when waiting for certs in the quickstart.  Not
strictly related to this change, but recent external factors have
caused increased delays at the same time we're tring to get this
merged.

Change-Id: I8c9315827e79d8a913b8c59889a18394e1671d33