Fix providers api endpoint when no images
We were incorrectly constructing the list of relevant images
and uploads for the provider endpoint; if there were no images
for a provider, then the build_artifacts and uploads variables
would be out of scope and throw an exception.

Instead we should accumulate relevant info for all images and
uploads or pass the empty list if there are none.

Change-Id: I3967f75317a230ae2a4e52ff1aad13bbf3288754