Allow ConnectionClosedError in unit test tracebacks
The change cache has watchers and does not have a shutdown/stop
method.  That means that it can continue to receive and process
events even after a test has completed.  If it does that and
performs a ZK get after the test framework has closed the zk
connection, it will log an exception.  This is harmless, but
trips the okay_tracebacks handler and fails the test.

Rather than add a stop/shutdown method to the change cache that
would only be used to avoid this issue in tests, let's instead
say that it's okay for any test to log the ConnectionClosedError.

It's possible that if we have real ZK problems with tests, this
may still happen, but either Zuul should recover or the test will
fail for legitimate reasons.  This isn't the sort of traceback
we need to be notified of if it occurs in tests.

Change-Id: Ibca8eb2c72d7c79a89c1cfaf3b9685df13a5e6c5