Support custom claim and print meta info in zuul-admin create-auth-token
This change includes two main changes:

1. Support custom claims:

When integrated with other auth service, e.g. Keycloak, zuul
authorization-rule could be configured  with claims in the token
from that service. For example, the token may contain claim like
"groups:tenant_x_admin", and it can be configured in the
conditions of the authroization-rule for the tenants, so that
users in that group can have access to the admin features on
zuul-web for tenant x.

Meanwhile, admin may want to reuse the same rule for API access
that use zuul-admin generated tokens, it can be achieved by
adding a custom claim "groups:teant_x_admin" to the token.

This can be useful for both admin-rules and access-rules.

This change supports the case by adding a new argument "--claim"
which can be used multiple times for adding multiple claims.

2. Print meta info of the token

In case the admin need to track each generated token, it would
be convienent by just copying the output meta info.

Change-Id: Ia69b15ff0b5d9de4e779a6d397a4d89af2ef68bd