Setup swap in nox jobs
Python3.12 appears to be more memory hungry than python3.11 leading to
test case failures when Ansible attempts to fork processes in the tests
leading to:

  ERROR! A worker was found in a dead state

We port openstack's configure-swap role into zuul so that we can
configure at least 1GB of swap on test nodes. The idea being while
slower this will prevent test cases from failing if they are on the edge
of running out of memory. A reasonable followup would be to inspect
where the memory consumption is going, but for now we're trying to
address the immediate problem of flaky jobs.

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/943861
Change-Id: I00f672ae20c8e8ddc3ac88d179b23271d7aa307b