Don't re-create job request on cancel/resume
With the job request queue operating mainly on cached requests we need
to consider that the cache could be slightly out-of-sync.

This means we need to consider that the request might no longer exist
when it's cancelled or resumed. So far we've used `ensure_path()` to
create those sub-znodes, which will also re-create missing request
nodes. The re-created request znode will however be empty and lead to
JSON decode errors.

Instead we'll assume the request znode still exists, try to create
cancel/resume nodes and ignore errors when the request was deleted in
the meantime or the sub-znodes already exists.

Change-Id: I574aa7425a270e7a6de26a003a0856e69ee451b8